+++ tests/test_utils_other.py
@@ -28,6 +28,9 @@ from datacube.utils.dates import date_sequence
 from datacube.utils.math import (
     num2numpy,
     is_almost_int,
+    maybe_zero,
+    maybe_int,
+    snap_scale,
     valid_mask,
     invalid_mask,
     clamp,
@@ -539,6 +542,31 @@ def test_is_almost_int():
     assert is_almost_int(1.001, .1)
     assert is_almost_int(2 - 0.001, .1)
     assert is_almost_int(-1.001, .1)
+
+
+def test_maybe_zero():
+    assert maybe_zero(0.0001, 0.1) == 0
+    assert maybe_zero(-0.0001, 0.1) == 0
+    assert maybe_zero(1.5, 0.1) == 1.5
+
+
+def test_maybe_int():
+    assert maybe_int(1, 1e-10) == 1
+    assert maybe_int(1.6, .1) == 1.6
+    assert maybe_int(-1.6, .1) == -1.6
+    assert maybe_int(1.001, .1) == 1
+    assert maybe_int(2 - 0.001, .1) == 2
+    assert maybe_int(-1.001, .1) == -1
+    assert maybe_int(1.1, .1) == 1.1
+
+
+def test_snap_scale():
+    assert snap_scale(0.9999999) == 1
+    assert snap_scale(-0.9999999) == -1
+    s = 0.999
+    assert snap_scale(0.999) is s
+    s = 0.621612621868
+    assert snap_scale(s) is s
 
 
 def test_valid_mask():