+++ docs/about/whats_new.rst
@@ -13,6 +13,7 @@ v1.8.4 (???)
 - Add ``cloud_cover`` to EO3 metadata
 - Add ``erosion`` functionality to Virtual products' ``ApplyMask`` to supplement existing ``dilation`` functionality (:pull:`1049`)
 - Fix numeric precision issues in ``compute_reproject_roi`` when pixel size is small. (:issue:`1047`)
+- Follow up fix to (:issue:`1047`) to round scale to nearest integer if very close.
 
 .. _`notebook examples`: https://github.com/GeoscienceAustralia/dea-notebooks/